MKL hello world done

+++ source/pythonInterface/matrixIntake.pyx
@@ -3,7 +3,7 @@ import numpy as np
 cimport numpy as np
 
 cdef extern void c_printMatrix(double *arr, int m, int n)
-cdef extern double c_dot(double *v, double *u, int n)
+cdef extern float c_dot(float *v, float *u, int n)
 
 @cython.boundscheck(False)
 @cython.wraparound(False)
@@ -14,7 +14,7 @@ def printMatrix(np.ndarray[double,ndim=2,mode="c"] A not None):
 	c_printMatrix(&A[0,0],m,n)
 	return None
 
-def customDot(np.ndarray[double,ndim=1,mode="c"] v not None,np.ndarray[double,ndim=1,mode="c"] u not None):
+def customDot(np.ndarray[float,ndim=1,mode="c"] v not None,np.ndarray[float,ndim=1,mode="c"] u not None):
 	cdef int n
 	n = v.shape[0]
 	assert(n == u.shape[0])
